requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt

celery
redis
channels
channels_redis
daphne
python-socketio
